+++ src/main/kotlin/racingcar/domain/CarPosition.kt
@@ -1,7 +1,9 @@
-package racingcar
+package racingcar.domain
 
 data class CarPosition(private val position: Int = 0) {
     fun move(): CarPosition {
         return CarPosition(position + 1)
     }
+
+    val point: Int get() = position
 }
